pyproject.toml
[project]
name = "gra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Otto Akseli Hannuksela", email = "otto.akseli.hannuksela@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "astropy>=7.2.0",
    "astroquery>=0.4.11",
    "async>=0.6.2",
    "bilby>=2.7.1",
    "gwosc>=0.8.1",
    "gwpy>=3.0.14",
    "ipython>=9.9.0",
    "lalsuite>=7.26.1",
    "ler>=0.5.9",
    "numpy>=2.4.1",
    "rich>=14.3.3",
    "scienceplots>=2.2.1",
    "typer-slim>=0.24.0",
    "zenodo-get>=3.0.0",
]

[project.scripts]
gra = "gra.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
